--- Content.Client/Rootable/RootableSystem.cs
@@ -1,5 +1,0 @@
-﻿using Content.Shared.Rootable;
-
-namespace Content.Client.Rootable;
-
-public sealed class RootableSystem : SharedRootableSystem;
